Address copilot pass 6 on #400
Profile-fetch cache extracted into the shared cid-doc-fetch.js so
the NIP-98 path benefits too. Previously only the LWS-CID verifier
had a TTL cache; the NIP-98 → WebID lookup did a fresh fetch on
every request, which both adds latency on the auth path and would
amplify self-traffic when the profile is hosted on the same server
that's serving the request.

Now both callers go through the same bounded LRU
(5-min hits / 1-min misses, 1000-entry cap, delete-then-set on hit
for LRU recency). The lws-cid.js side reduces to a tiny delegate;
the cache + the network primitive live in one module.

Tests:
- _clearProfileCacheForTests is now exported from cid-doc-fetch.js
  (re-exported from lws-cid.js for back-compat with its existing
  test). nostr-cid-vm.test.js imports it directly and clears in
  beforeEach so a previous test's cached profile can't satisfy a
  later test (since every test rotates the privkey).
- Both modules' tests still pass (47 + 15 = 62) and full suite
  goes 699 → 700 with no regressions.
